+++ svelte.config.js
@@ -14,6 +14,12 @@ const config = {
 		prerender: {
 			entries: ['/', '*']
 		}
+	},
+
+	compilerOptions: {
+		experimental: {
+			async: true
+		}
 	}
 };