+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - new export: @hyvor/design/cloud
 - HyvorBar moved to cloud
 - CloudContext is now required for Consoles
+- CodeBlock now uses Shiki.js instead of Highlight.js
+- CodeBlock requires `compilerOptions.experimental.async = true` in `svelte.config.js` (https://svelte.dev/docs/svelte/await-expressions)
 
 ## 1.1.22 - 2025-11-05